@@ -5,8 +5,11 @@ import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.support.v4.app.FragmentManager;
+import androidx.fragment.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
+import androidx.fragment.app.FragmentTransaction;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -22,7 +25,6 @@ import com.liftdom.liftdom.R;
  * A simple {@link Fragment} subclass.
  */
 public class KnowledgeCenterMainFrag extends Fragment {
-
 
     public KnowledgeCenterMainFrag() {
         // Required empty public constructor
